+++ tests/unit.lua
@@ -76,6 +76,7 @@ dofile('./tests/Models/PlayerTest.lua')
 dofile('./tests/Support/ArrTest.lua')
 dofile('./tests/Support/BoolTest.lua')
 dofile('./tests/Support/StrTest.lua')
+dofile('./tests/Support/IntervalTest.lua')
 
 dofile('./tests/Views/ViewConstantsTest.lua')
 dofile('./tests/Views/Windows/WindowTest.lua')